Found the bug cause
Keypoints are in local coordinate system! This means that when we add them to  self._train_kp  they refer to the original image, and not to the patch in the mosaic. Before adding them to self._train_kp, we must transform their  .pt  with the  H_img_to_mosaic  we obtained after the matching process!

[WIP] algorithm breaks with more than 2 images

[BUG] images must have enough common keypoints
[WIP] custom warping operation for the SPHP homography method